learning objectives
As per Jolene's question and Courtney's response, I figured it would be wise to submit PRs early and often. For these learning objectives, I think the changes I'm suggesting will make it more clear to the teacher and student what is expected of them. I may have erred somewhat in the scope of this intro section, please correct where necessary. However, I think focusing on what the participants should be able to do by the end of this section will help with the next round of revisions in later parts of this intro and in the rest of the curriculum.